+++ pkg/clients/mocks/ServerClient.go
@@ -200,6 +200,36 @@ func (_m *ServerClient) GetApplicationCode(_a0 context.Context, _a1 *time.Durati
 	return r0, r1
 }
 
+// GetDatabaseLifecyclePhysicalDatabaseInstance provides a mock function with given fields: _a0, _a1, _a2, _a3
+func (_m *ServerClient) GetDatabaseLifecyclePhysicalDatabaseInstance(_a0 context.Context, _a1 *time.Duration, _a2 http.Header, _a3 string) (*http.Response, error) {
+	ret := _m.Called(_a0, _a1, _a2, _a3)
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetDatabaseLifecyclePhysicalDatabaseInstance")
+	}
+
+	var r0 *http.Response
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *time.Duration, http.Header, string) (*http.Response, error)); ok {
+		return rf(_a0, _a1, _a2, _a3)
+	}
+	if rf, ok := ret.Get(0).(func(context.Context, *time.Duration, http.Header, string) *http.Response); ok {
+		r0 = rf(_a0, _a1, _a2, _a3)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*http.Response)
+		}
+	}
+
+	if rf, ok := ret.Get(1).(func(context.Context, *time.Duration, http.Header, string) error); ok {
+		r1 = rf(_a0, _a1, _a2, _a3)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
 // GetDatabaseLifecyclePhysicalDatabaseInstances provides a mock function with given fields: _a0, _a1, _a2, _a3
 func (_m *ServerClient) GetDatabaseLifecyclePhysicalDatabaseInstances(_a0 context.Context, _a1 *time.Duration, _a2 http.Header, _a3 clients.DatabaseLifecyclePhysicalDatabaseInstanceListRequest) (*http.Response, error) {
 	ret := _m.Called(_a0, _a1, _a2, _a3)